+++ cdk/kraken/src/docker.ts
@@ -1,4 +1,4 @@
-import { CheckoutJob, Workflow, JobProps } from 'cdkactions';
+import { CheckoutJob, Workflow, JobProps, StepsProps } from 'cdkactions';
 import { buildId, buildName } from './utils';
 
 
@@ -52,13 +52,24 @@ export interface DockerPublishJobProps {
    * @default true
    */
   cache?: boolean;
+
+  /**
+   * If enabled, do not publish docker image to Docker Hub.
+   * Instead upload as an artifact.
+   * @default false
+   */
+  noPublish?: boolean;
 }
 
 /**
  * A job to build and publish a Docker image.
- * TODO: look into docker build action v2
  */
 export class DockerPublishJob extends CheckoutJob {
+  /**
+   * Full name of the docker image built.
+   */
+  public readonly dockerImageName: string;
+
   /**
    *
    * @param scope cdkactions Workflow instance.
@@ -76,32 +87,81 @@ export class DockerPublishJob extends CheckoutJob {
       password: '${{ secrets.DOCKER_PASSWORD }}',
       dockerfile: 'Dockerfile',
       cache: true,
+      noPublish: false,
       ...config,
     };
+    const formattedName = fullConfig.noPublish ? buildName('Build', id) : buildName('Publish', id);
+    const formattedId = fullConfig.noPublish ? buildId('build', id) : buildId('publish', id);
     // Define docker action with
+    const imageName = `pennlabs/${fullConfig.imageName}`;
     const dockerWith: any = {
-      repository: `pennlabs/${fullConfig.imageName}`,
-      path: fullConfig.path,
-      dockerfile: `${fullConfig.path}/${fullConfig.dockerfile}`,
-      username: fullConfig.username,
-      password: fullConfig.password,
-      push: fullConfig.push,
-      tags: fullConfig.tags,
+      'context': fullConfig.path,
+      'file': `${fullConfig.path}/${fullConfig.dockerfile}`,
+      'push': fullConfig.push,
+      'cache-from': 'type=local,src=/tmp/.buildx-cache',
+      'cache-to': 'type=local,dest=/tmp/.buildx-cache',
     };
     if (fullConfig.cache) {
-      dockerWith.cache_froms = `${dockerWith.repository}:latest`;
+      dockerWith['cache-from'] = dockerWith['cache-from'].concat(`,type=registry,ref=${imageName}:latest`);
+    }
+
+    // Build tags string
+    const tagsString = (tags: string) => tags.split(',').map(tag => `${imageName}:${tag}`).join();
+    dockerWith.tags = tagsString(fullConfig.tags);
+
+    // Define steps
+    const steps: StepsProps[] = [
+      {
+        uses: 'docker/setup-qemu-action@v1',
+      },
+      {
+        uses: 'docker/setup-buildx-action@v1',
+      },
+      {
+        name: 'Cache Docker layers',
+        uses: 'actions/cache@v2',
+        with: {
+          path: '/tmp/.buildx-cache',
+          key: `buildx-${formattedId}`,
+        },
+      },
+      {
+        uses: 'docker/login-action@v1',
+        with: {
+          username: fullConfig.username,
+          password: fullConfig.password,
+        },
+      },
+      {
+        name: 'Build/Publish',
+        uses: 'docker/build-push-action@v2',
+        with: dockerWith,
+      },
+    ];
+
+    // If publishing isn't wanted
+    if (fullConfig.noPublish) {
+      dockerWith.push = false;
+      dockerWith.outputs = 'type=docker,dest=/tmp/image.tar';
+      steps.push({
+        uses: 'actions/upload-artifact@v2',
+        with: {
+          name: formattedId,
+          path: '/tmp/image.tar',
+        },
+      });
+      steps.splice(3, 1);
     }
 
     // Create job
-    super(scope, buildId('publish', id), {
-      name: buildName('Publish', id),
+    super(scope, formattedId, {
+      name: formattedName,
       runsOn: 'ubuntu-latest',
-      steps: [{
-        name: 'Publish',
-        uses: 'docker/build-push-action@v1',
-        with: dockerWith,
-      }],
+      steps,
       ...overrides,
     });
+
+    // Set public fields
+    this.dockerImageName = imageName;
   }
 }