+++ cdk/kraken/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.4.5 (2021-01-30)
+
+* Update `docker/build-push-action` to v2
+* Complete integration test job
+
 ## 0.3.10 (2020-12-30)
 
 * Initial beta release